requirements.txt
requests
beautifulsoup4
playwright
pytz
google-genai